refactor: changed pages names to index.tsx
